Fix the unsigned signed compare issue

+++ tests/mock_tests/check.h
@@ -56,8 +56,8 @@ struct Check
             auto act_len = sai_serialize_attribute_value(&act_str[0], meta, &act_attr_list[i].value);
             auto exp_len = sai_serialize_attribute_value(&exp_str[0], meta, &exp_attr_list.get_attr_list()[i].value);
 
-            assert(act_len < act_str.size());
-            assert(act_len < exp_str.size());
+            assert(act_len < (int)act_str.size());
+            assert(act_len < (int)exp_str.size());
 
             if (act_len != exp_len)
             {
